refactor: Reference current state using ushort instead of state itself.
BREAKING CHANGE: Changed the state identifier from a int to a ushort.
CurrentState now refered to by it's number instead of saving a reference to it.
Removed changing state by providing a state directly.

+++ Samples/TDAction/Scripts/Entities/Managers/EntityCombatManager.cs
@@ -87,7 +87,7 @@ namespace TDAction.Entities
             }
             else
             {
-                manager.StateManager.ChangeState((int)(manager.IsGrounded ? EntityStates.FLINCH_GROUND : EntityStates.FLINCH_AIR));
+                manager.StateManager.ChangeState((ushort)(manager.IsGrounded ? EntityStates.FLINCH_GROUND : EntityStates.FLINCH_AIR));
             }
             return hitReaction;
         }
